Fixup: Add tests
