+++ l10n/bundle.l10n.json
@@ -141,9 +141,9 @@
   "Text editor must be focused to fix all issues": "Text editor must be focused to fix all issues",
   "Fix all issues": "Fix all issues",
   "Select fix all action": "Select fix all action",
-  "Packaged Copilot LSP config is missing lspServers.csharp.": "Packaged Copilot LSP config is missing lspServers.csharp.",
   "Failed to read packaged Copilot LSP config: {0}": "Failed to read packaged Copilot LSP config: {0}",
   "Failed to read Copilot LSP config: {0}": "Failed to read Copilot LSP config: {0}",
+  "Failed to update Copilot LSP config: {0}": "Failed to update Copilot LSP config: {0}",
   "Copilot LSP config already contains roslyn-language-server. No changes were made.": "Copilot LSP config already contains roslyn-language-server. No changes were made.",
   "Updated Copilot LSP config at {0}.": "Updated Copilot LSP config at {0}.",
   "Failed to write Copilot LSP config: {0}": "Failed to write Copilot LSP config: {0}",
